requirements.txt
flask
requests
pygal
python-dotenv